Added warning that master-slave communication unoperational
Currently if you try and hook up a strip of RGBs directly to the slave side and the other strip directly to the master, the slave side's won't update. I added mention of that fact to ensure that someone doesn't build a board expecting that to work, like I did.